--- api/__init__.py
@@ -1,3 +1,0 @@
-"""
-API package for Foodgram project.
-"""
