+++ api/apps.py
@@ -1,6 +1,3 @@
-"""
-API app configuration.
-"""
 from django.apps import AppConfig
 
 class ApiConfig(AppConfig):
