--- api/recipes/__init__.py
@@ -1,3 +1,0 @@
-"""
-Recipes API package.
-"""
