+++ api/recipes/serializers.py
@@ -1,8 +1,7 @@
-"""
-Serializers for the recipes API.
-"""
 from rest_framework import serializers
-from recipes.models import Recipe, Ingredient, Tag
+from recipes.models import Recipe, Ingredient, Tag, RecipeIngredient, Favorite, ShoppingCart
+from recipes.fields import Base64ImageField
+from api.users.serializers import CustomUserSerializer
 
 class IngredientSerializer(serializers.ModelSerializer):
     """Serializer for ingredients."""
@@ -20,7 +19,75 @@ class RecipeSerializer(serializers.ModelSerializer):
     """Serializer for recipes."""
     ingredients = IngredientSerializer(many=True, read_only=True)
     tags = TagSerializer(many=True, read_only=True)
+    image = Base64ImageField()
+    is_favorited = serializers.SerializerMethodField()
+    is_in_shopping_cart = serializers.SerializerMethodField()
+    author = CustomUserSerializer(read_only=True)
     
     class Meta:
         model = Recipe
-        fields = ['id', 'name', 'text', 'cooking_time', 'ingredients', 'tags'] 
+        fields = [
+            'id', 'tags', 'author', 'ingredients', 'is_favorited',
+            'is_in_shopping_cart', 'name', 'image', 'text', 'cooking_time'
+        ]
+
+    def get_is_favorited(self, obj):
+        request = self.context.get('request')
+        if not request or request.user.is_anonymous:
+            return False
+        return Favorite.objects.filter(user=request.user, recipe=obj).exists()
+
+    def get_is_in_shopping_cart(self, obj):
+        request = self.context.get('request')
+        if not request or request.user.is_anonymous:
+            return False
+        return ShoppingCart.objects.filter(user=request.user, recipe=obj).exists()
+
+class RecipeCreateSerializer(serializers.ModelSerializer):
+    """Serializer for creating recipes."""
+    ingredients = serializers.ListField(
+        child=serializers.DictField(child=serializers.IntegerField())
+    )
+    tags = serializers.PrimaryKeyRelatedField(
+        queryset=Tag.objects.all(), many=True
+    )
+    image = Base64ImageField()
+
+    class Meta:
+        model = Recipe
+        fields = [
+            'ingredients', 'tags', 'image', 'name', 'text', 'cooking_time'
+        ]
+
+    def to_representation(self, instance):
+        """Convert instance to dictionary for response."""
+        return RecipeSerializer(instance, context=self.context).data
+
+    def create(self, validated_data):
+        ingredients_data = validated_data.pop('ingredients')
+        tags_data = validated_data.pop('tags')
+        recipe = Recipe.objects.create(**validated_data)
+        recipe.tags.set(tags_data)
+
+        for ingredient_data in ingredients_data:
+            RecipeIngredient.objects.create(
+                recipe=recipe,
+                ingredient_id=ingredient_data.get('id'),
+                amount=ingredient_data.get('amount'),
+            )
+        return recipe
+
+    def update(self, instance, validated_data):
+        if 'ingredients' in validated_data:
+            ingredients_data = validated_data.pop('ingredients')
+            instance.ingredients.clear()
+            for ingredient_data in ingredients_data:
+                RecipeIngredient.objects.create(
+                    recipe=instance,
+                    ingredient_id=ingredient_data.get('id'),
+                    amount=ingredient_data.get('amount'),
+                )
+        if 'tags' in validated_data:
+            tags_data = validated_data.pop('tags')
+            instance.tags.set(tags_data)
+        return super().update(instance, validated_data)
