+++ recipes/apps.py
@@ -1,6 +1,3 @@
-"""
-Recipes app configuration.
-"""
 from django.apps import AppConfig
