--- recipes/tests.py
@@ -1,2 +1,0 @@
-
-# Create your tests here.
