--- users/__init__.py
@@ -1,3 +1,0 @@
-"""
-Users app for Foodgram project.
-"""
