+++ users/apps.py
@@ -1,6 +1,3 @@
-"""
-Users app configuration.
-"""
 from django.apps import AppConfig
 
 class UsersConfig(AppConfig):
