+++ users/models.py
@@ -1,6 +1,3 @@
-"""
-Models for the users app.
-"""
 from django.contrib.auth.models import AbstractUser
 from django.db import models
 from django.db.models import UniqueConstraint
